Move the EvaluatorModelView methods in the derived classes
