+++ force_wfmanager/left_side_pane/data_source_model_view.py
@@ -1,4 +1,4 @@
-from traits.api import Instance
+from traits.api import Instance, on_trait_change
 
 from force_bdss.api import BaseDataSourceModel
 
@@ -8,3 +8,14 @@ from .evaluator_model_view import EvaluatorModelView
 class DataSourceModelView(EvaluatorModelView):
     #: DataSource model (More restrictive than the ModelView model attribute)
     model = Instance(BaseDataSourceModel, allow_none=False)
+
+    def __evaluator_default(self):
+        return self.model.factory.create_data_source()
+
+    @on_trait_change(
+        'variable_names_registry.data_source_available_variables[]')
+    def update_data_source_input_rows(self):
+        available_variables = \
+            self.variable_names_registry.data_source_available_variables
+        for input_slot_row in self.input_slots_representation:
+            input_slot_row.available_variables = available_variables
