+++ force_wfmanager/left_side_pane/kpi_calculator_model_view.py
@@ -1,4 +1,4 @@
-from traits.api import Instance
+from traits.api import Instance, on_trait_change
 
 from force_bdss.api import BaseKPICalculatorModel
 
@@ -9,3 +9,14 @@ class KPICalculatorModelView(EvaluatorModelView):
     #: KPI Calculator model (More restrictive than the ModelView model
     #: attribute)
     model = Instance(BaseKPICalculatorModel, allow_none=False)
+
+    def __evaluator_default(self):
+        return self.model.factory.create_kpi_calculator()
+
+    @on_trait_change(
+        'variable_names_registry.kpi_calculator_available_variables[]')
+    def update_kpi_calculator_input_rows(self):
+        available_variables = \
+            self.variable_names_registry.kpi_calculator_available_variables
+        for input_slot_row in self.input_slots_representation:
+            input_slot_row.available_variables = available_variables
